requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
requests==2.31.0
python-magic==0.4.27
gunicorn==21.2.0
starlette==0.27.0
click==8.2.1
h11==0.16.0
email-validator==2.1.0.post1
together==0.1.2
python-dotenv==1.0.0
